synonyms are working for commands, added image component, return undefinded if command not valid

+++ src/api/services/buildHtml.js
@@ -20,18 +20,18 @@ function buildHtml(buildedCommand) {
   template = template.filter(function (el) {
     return el != null
   })
-
   if (buildedCommand.action == 'add') {
     let component = components[buildedCommand.element]
-
+    if (!component) return undefined
     component['id'] = findFreeId(template)
     if (buildedCommand.id) {
       addIntoElementWithId(template, component, buildedCommand.id)
     } else template.push(component)
   } else if (buildedCommand.action == 'delete') {
     //find by id and delete this element
+    if (!buildedCommand.id) return undefined
     deleteById(template, buildedCommand.id)
-  }
+  } else return undefined
 
   let html = json2html.render({}, template)
 
@@ -57,10 +57,9 @@ function findFreeId(template) {
 
 function addIntoElementWithId(template, component, id) {
   for (const key of Object.keys(template)) {
-    if (template[key]['id'] == id)
-      template[key]['html'].push(component)
+    if (template[key]['id'] == id) template[key]['html'].push(component)
     else if (Array.isArray(template[key]['html']))
-    addIntoElementWithId(template[key], component, id)
+      addIntoElementWithId(template[key], component, id)
   }
 }
